@@ -14,13 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+using SoulMemory.Native;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
-using SoulMemory.Native;
 
 namespace SoulMemory.soulmods;
 
@@ -58,13 +58,25 @@ public static class Soulmods
         var soulmodsModuleName = process.SoulmodsModuleName();
         var path = Path.Combine(dir, soulmodsModuleName);
 
-        process.InjectDll(path);
-
-        if (process.Modules.Cast<ProcessModule>().Any(processModule => processModule.ModuleName == soulmodsModuleName))
+        //potentially its already injecting, if this is the case we re-use it
+        var modules = process.GetModulesViaSnapshot();
+        if (modules.Any(processModule => processModule.szModule == soulmodsModuleName))
         {
             _soulmodsExports = process.GetModuleExports(soulmodsModuleName);
             return Result.Ok(_soulmodsExports);
         }
+
+        //If not, inject
+        process.InjectDll(path);
+
+        //Make sure it is in the module list now
+        modules = process.GetModulesViaSnapshot();
+        if (modules.Any(processModule => processModule.szModule == soulmodsModuleName))
+        {
+            _soulmodsExports = process.GetModuleExports(soulmodsModuleName);
+            return Result.Ok(_soulmodsExports);
+        }
+
         return Result.Err();
     }
 
