@@ -14,107 +14,81 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+using SoulMemory.Native;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Windows.Forms;
 using System.Windows.Input;
-using SoulSplitter.Native;
 
-namespace SoulSplitter.Hotkeys;
-
-public static class GlobalHotKey
+namespace SoulSplitter.Hotkeys
 {
-    private static readonly List<(int id, ModifierKeys modifier, Key key, Action action)> Hotkeys = [];
-    private static readonly ManualResetEvent WindowReadyEvent = new(false);
-    public static volatile HotkeyForm? HotkeyForm;
-    public static volatile IntPtr Handle;
-    private static int _currentId;
-
-    static GlobalHotKey()
+    public static class GlobalHotkey
     {
-        var messageLoopThread = new Thread(delegate ()
-        {
-            Application.Run(new HotkeyForm(WindowReadyEvent));
-        });
-        messageLoopThread.Name = "MessageLoopThread";
-        messageLoopThread.IsBackground = true;
-        messageLoopThread.Start();
-    }
+        private static readonly List<(int id, Hotkey hotkey, Action action)> Hotkeys = new List<(int, Hotkey, Action)>();
+        private static readonly ManualResetEvent WindowReadyEvent = new ManualResetEvent(false);
+        public static volatile HotkeyForm HotkeyForm = null!;
+        public static volatile IntPtr Handle;
+        private static int _currentId;
 
-    public static void OnHotkeyPressed(ModifierKeys modifier, Key key)
-    {
-        Hotkeys.ForEach(x =>
+        static GlobalHotkey()
         {
-            if (modifier == x.modifier && key == x.key)
+            var messageLoopThread = new Thread(delegate ()
             {
-                x.action();
+                Application.Run(new HotkeyForm(WindowReadyEvent));
+            });
+            messageLoopThread.Name = "MessageLoopThread";
+            messageLoopThread.IsBackground = true;
+            messageLoopThread.Start();
+        }
+
+        public static void OnHotkeyPressed(Hotkey hotkey)
+        {
+            Hotkeys.ForEach(x =>
+            {
+                if (hotkey.Modifiers == x.hotkey.Modifiers && hotkey.Key == x.hotkey.Key)
+                {
+                    x.action();
+                }
+            });
+        }
+
+        public static int RegisterHotKey(Hotkey hotkey, Action action)
+        {
+            WindowReadyEvent.WaitOne(); //wait for hotkey window to have initialized
+
+            var virtualKeyCode = (Keys)KeyInterop.VirtualKeyFromKey(hotkey.Key);
+            int id = Interlocked.Increment(ref _currentId);
+
+            Delegate register = (Action)(() =>
+            {
+                User32.RegisterHotkey(
+                    Handle,
+                    id,
+                    (uint)hotkey.Modifiers | 0x4000, //no repeat
+                    (uint)virtualKeyCode);
+            });
+
+            HotkeyForm.Invoke(register);
+            Hotkeys.Add((id, hotkey, action));
+            return id;
+        }
+
+        public static void UnregisterHotKey(int id)
+        {
+            if (Hotkeys.All(i => i.id != id))
+            {
+                return;
             }
-        });
-    }
 
-    public static int RegisterHotKey(ModifierKeys modifier, Key key, Action action)
-    {
-        WindowReadyEvent.WaitOne(); //wait for hotkey window to have initialized
+            Delegate unregister = (Action)(() =>
+            {
+                User32.UnregisterHotkey(Handle, id);
+            });
 
-        var virtualKeyCode = (Keys)KeyInterop.VirtualKeyFromKey(key);
-        var id = Interlocked.Increment(ref _currentId);
-
-        Delegate register = (Action)(() =>
-        {
-            User32.RegisterHotkey(
-                Handle,
-                id,
-                (uint)modifier | 0x4000, //no repeat
-                (uint)virtualKeyCode);
-        });
-
-        HotkeyForm?.Invoke(register);
-        Hotkeys.Add((id, modifier, key, action));
-        return id;
-    }
-
-    public static void UnregisterHotKey(int id)
-    {
-        if (Hotkeys.All(i => i.id != id))
-        {
-            return;
+            HotkeyForm.Invoke(unregister);
+            Hotkeys.Remove(Hotkeys.Find(i => i.id == id));
         }
-
-        Delegate unregister = (Action)(() =>
-        {
-            User32.UnregisterHotkey(Handle, id);
-        });
-
-        HotkeyForm?.Invoke(unregister);
-        Hotkeys.Remove(Hotkeys.Find(i => i.id == id));
-    }
-}
-
-public class HotkeyForm : Form
-{
-    public HotkeyForm(ManualResetEvent windowReadyEvent)
-    {
-        GlobalHotKey.Handle = Handle;
-        GlobalHotKey.HotkeyForm = this;
-        windowReadyEvent.Set();
-    }
-
-    protected override void WndProc(ref Message windowMessage)
-    {
-        base.WndProc(ref windowMessage);
-
-        if (windowMessage.Msg == 0x0312) //0x0312 is the hotkey message
-        {
-            var key = KeyInterop.KeyFromVirtualKey((int)windowMessage.LParam >> 16 & 0xFFFF);
-            var modifier = (ModifierKeys)((int)windowMessage.LParam & 0xFFFF);
-            GlobalHotKey.OnHotkeyPressed(modifier, key);
-        }
-    }
-
-    protected override void SetVisibleCore(bool value)
-    {
-        base.SetVisibleCore(false);
     }
 }
