@@ -14,6 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+
 use std::sync::{Arc, Mutex};
 use windows::Win32::Foundation::HINSTANCE;
 use imgui::{Condition, Context, Ui};
@@ -25,6 +26,7 @@ use crate::widgets::chr_dbg_flags_widget::ChrDbgFlagsWidget;
 use crate::widgets::event_flag_widget::EventFlagWidget;
 use crate::widgets::misc_widget::MiscWidget;
 use crate::games::*;
+use crate::util::WINDOW_TITLE_STR;
 use crate::widgets::emevd_logger_widget::EmevdLoggerWidget;
 
 pub struct App
@@ -65,21 +67,35 @@ impl App
     pub fn new(process_name: &String, hmodule: HINSTANCE) -> Self
     {
         //Init the game we're injected in
+
+        #[cfg(target_arch = "x86")]
         let game: Box<dyn Game> = match process_name.to_lowercase().as_str()
         {
             "mockgame.exe"              => Box::new(MockGame::new()),
+
             "darksouls.exe"             => Box::new(DarkSoulsPrepareToDieEdition::new()),
-            #[cfg(target_arch = "x86_64")]
-            "darksoulsii.exe"           => Box::new(DarkSouls2ScholarOfTheFirstSin::new()),
-            #[cfg(target_arch = "x86")]
             "darksoulsii.exe"           => Box::new(DarkSouls2Vanilla::new()),
+
+            _                           => panic!("unsupported process: {}", process_name.to_lowercase()),
+        };
+
+        #[cfg(target_arch = "x86_64")]
+        let game: Box<dyn Game> = match process_name.to_lowercase().as_str()
+        {
+            "mockgame.exe"              => Box::new(MockGame::new()),
+
             "darksoulsremastered.exe"   => Box::new(DarkSoulsRemastered::new()),
+            "darksoulsii.exe"           => Box::new(DarkSouls2ScholarOfTheFirstSin::new()),
             "darksoulsiii.exe"          => Box::new(DarkSouls3::new()),
             "sekiro.exe"                => Box::new(Sekiro::new()),
             "eldenring.exe"             => Box::new(EldenRing::new()),
             "armoredcore6.exe"          => Box::new(ArmoredCore6::new()),
+            "shadps4.exe"               => Box::new(Bloodborne::new()),
+            "nightreign.exe"            => Box::new(Nightreign::new()),
+
             _                           => panic!("unsupported process: {}", process_name.to_lowercase()),
         };
+
 
         //get drawable widgets
         //let widgets = game.get_widgets();
@@ -195,9 +211,10 @@ impl App
         //style.tab_rounding                            = 4f32;
     }
 
+
     pub fn render(&mut self, ui: &mut Ui)
     {
-        ui.window("soulmemory-rs")
+        ui.window(WINDOW_TITLE_STR)
             //.flags(WindowFlags::NO_COLLAPSE | WindowFlags::NO_SCROLLBAR)
             .position([50.0f32, 50.0f32], Condition::Appearing)
             .size([500.0, 800.0], Condition::Appearing)
@@ -230,3 +247,4 @@ impl Default for App
 }
 
 static mut APP: Option<Arc<Mutex<App>>> = None;
+
