@@ -1,8 +1,24 @@
+// This file is part of the SoulSplitter distribution (https://github.com/FrankvdStam/SoulSplitter).
+// Copyright (c) 2022 Frank van der Stam.
+// https://github.com/FrankvdStam/SoulSplitter/blob/main/LICENSE
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, version 3.
+//
+// This program is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
 use std::{fmt, mem};
 use std::fmt::Display;
 use std::sync::{Arc, Mutex};
 use chrono::{DateTime, Local};
-use crate::darkscript3::sekiro_emedf::Emedf;
+use crate::darkscript3::emevd_definition::EmevdDefinition;
 
 pub struct BufferedEmevdCall
 {
@@ -46,5 +62,5 @@ pub trait BufferedEmevdLogger
         return buffer;
     }
 
-    fn get_game_emevd_definitions(&self) -> &Emedf;
+    fn get_game_emevd_definitions(&self) -> &EmevdDefinition;
 }
