@@ -14,17 +14,26 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+
 pub(crate) mod log;
 pub(crate) mod console;
 pub(crate) mod server;
 pub mod vector3f;
 
+#[unsafe(no_mangle)]
+#[used]
+pub static SOULMEMORY_RS_VERSION: &'static str = env!("VERSION");
+pub const WINDOW_TITLE_STR: &'static str = concat!("soulmemory-rs ", env!("VERSION"));
+
+
+#[allow(dead_code)]
 pub unsafe fn get_stack_u32(esp: u32, offset: usize) -> u32
 {
-    *((esp as usize + offset) as usize as *mut u32)
+    unsafe { *((esp as usize + offset) as *mut u32) }
 }
 
+#[allow(dead_code)]
 pub unsafe fn get_stack_u8(esp: u32, offset: usize) -> u8
 {
-    *((esp as usize + offset) as usize as *mut u8)
+    unsafe {  *((esp as usize + offset) as *mut u8) }
 }
