@@ -14,12 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-use std::{thread, time::Duration};
+use std::{thread, time::Duration, mem};
 
 use ilhook::x64::{Hooker, HookType, Registers, CallbackOption, HookFlags, HookPoint};
 use mem_rs::prelude::*;
 
 use log::info;
+
+use windows::Win32::UI::Input::XboxController::*;
 
 use crate::util::GLOBAL_VERSION;
 use crate::util::Version;
@@ -32,13 +34,14 @@ struct FpsOffsets
     timestamp_current: isize,
 }
 
-static mut IGT_BUFFER: f32 = 0.0f32;
+pub(crate) static mut IGT_BUFFER: f32 = 0.0f32;
 static mut IGT_HOOK: Option<HookPoint> = None;
 
 static mut FPS_HOOK: Option<HookPoint> = None;
 static mut FPS_HISTORY_HOOK: Option<HookPoint> = None;
 static mut FPS_CUSTOM_LIMIT_HOOK: Option<HookPoint> = None;
 static mut FRAME_ADVANCE_HOOK: Option<HookPoint> = None;
+static mut XINPUT_HOOK: Option<HookPoint> = None;
 
 static mut FPS_OFFSETS: FpsOffsets = FpsOffsets {
     target_frame_delta: 0x0,
@@ -47,21 +50,43 @@ static mut FPS_OFFSETS: FpsOffsets = FpsOffsets {
     timestamp_current: 0x0,
 };
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 #[used]
 pub static mut ER_FPS_PATCH_ENABLED: bool = false;
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 #[used]
 pub static mut ER_FPS_CUSTOM_LIMIT: f32 = 0.0f32;
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 #[used]
 pub static mut ER_FRAME_ADVANCE_ENABLED: bool = false;
 
-#[no_mangle]
+#[unsafe(no_mangle)]
 #[used]
 pub static mut ER_FRAME_RUNNING: bool = false;
+
+#[unsafe(no_mangle)]
+#[used]
+pub static mut ER_XINPUT_PATCH_ENABLED: bool = false;
+
+#[unsafe(no_mangle)]
+#[used]
+pub static mut ER_XINPUT_STATE: XINPUT_STATE = XINPUT_STATE {
+    dwPacketNumber: 0,
+    Gamepad: XINPUT_GAMEPAD {
+        wButtons: XINPUT_GAMEPAD_BUTTON_FLAGS(0),
+        bLeftTrigger: 0,
+        bRightTrigger: 0,
+        sThumbLX: 0,
+        sThumbLY: 0,
+        sThumbRX: 0,
+        sThumbRY: 0,
+    }
+};
+
+
+pub type XInputGetState = unsafe extern "system" fn(dw_user_index: u32, p_state: *mut XINPUT_STATE) -> u32;
 
 
 #[allow(unused_assignments)]
@@ -72,7 +97,7 @@ pub fn init_eldenring()
         info!("version: {}", GLOBAL_VERSION);
         
         // Get ER process
-        let mut process = Process::new("eldenring.exe");
+        let mut process = Process::new_with_memory_type("eldenring.exe", MemoryType::Direct);
         process.refresh().unwrap();
 
         // AoB scan for timer patch
@@ -80,7 +105,7 @@ pub fn init_eldenring()
         info!("increment IGT at 0x{:x}", fn_increment_igt_address);
 
         // Enable timer patch
-        IGT_HOOK = Some(Hooker::new(fn_increment_igt_address, HookType::JmpBack(increment_igt), CallbackOption::None, 0, HookFlags::empty()).hook().unwrap());
+        IGT_HOOK = Some(Hooker::new(fn_increment_igt_address, HookType::JmpBack(increment_igt_hook), CallbackOption::None, 0, HookFlags::empty()).hook().unwrap());
 
         // AoBs for FPS patches
         let mut fps_aob = "";
@@ -148,28 +173,40 @@ pub fn init_eldenring()
 
         // Enable frame advance patch
         FRAME_ADVANCE_HOOK = Some(Hooker::new(fn_frame_advance_address, HookType::JmpBack(frame_advance), CallbackOption::None, 0, HookFlags::empty()).hook().unwrap());
+
+
+        // Find XInputGetState function in XINPUT1_4.dll
+        let xinput_module = process.get_modules().iter().find(|m| m.name == "XINPUT1_4.dll").cloned().expect("Couldn't find XINPUT1_4.dll");
+        let xinput_fn_addr = xinput_module.get_exports().iter().find(|e| e.0 == "XInputGetState").expect("Couldn't find XInputGetState").1;
+        info!("XInputGetState at 0x{:x}", xinput_fn_addr);
+
+        // Hook XInputGetState
+        XINPUT_HOOK = Some(Hooker::new(xinput_fn_addr, HookType::Retn(xinput_fn), CallbackOption::None, 0, HookFlags::empty()).hook().unwrap());
     }
 }
 
-unsafe extern "win64" fn increment_igt(registers: *mut Registers, _:usize)
+pub unsafe extern "win64" fn increment_igt_hook(registers: *mut Registers, _:usize)
 {
-    let mut frame_delta = std::mem::transmute::<u32, f32>((*registers).xmm0 as u32);
-    //convert to milliseconds
-    frame_delta = frame_delta * 1000f32;
-    frame_delta = frame_delta * 0.96f32; //scale to IGT
-
-    //Rather than casting, like the game does, make the behavior explicit by flooring
-    let mut floored_frame_delta = frame_delta.floor();
-    let remainder = frame_delta - floored_frame_delta;
-    IGT_BUFFER = IGT_BUFFER + remainder;
-
-    if IGT_BUFFER > 1.0f32
+    unsafe
     {
-        IGT_BUFFER = IGT_BUFFER - 1f32;
-        floored_frame_delta += 1f32;
-    }
+        let mut frame_delta = f32::from_bits((*registers).xmm0 as u32);
+        //convert to milliseconds
+        frame_delta = frame_delta * 1000f32;
+        frame_delta = frame_delta * 0.96f32; //scale to IGT
 
-    (*registers).xmm1 = std::mem::transmute::<f32, u32>(floored_frame_delta) as u128;
+        //Rather than casting, like the game does, make the behavior explicit by flooring
+        let mut floored_frame_delta = frame_delta.floor();
+        let remainder = frame_delta - floored_frame_delta;
+        IGT_BUFFER = IGT_BUFFER + remainder;
+
+        if IGT_BUFFER > 1.0f32
+        {
+            IGT_BUFFER = IGT_BUFFER - 1f32;
+            floored_frame_delta += 1f32;
+        }
+        //println!("ER hook before: {} scaled: {} buffer {} after {}", before, frame_delta, IGT_BUFFER, floored_frame_delta);
+        (*registers).xmm1 = f32::to_bits(floored_frame_delta) as u128;
+    }
 }
 
 // FPS patch
@@ -179,27 +216,30 @@ unsafe extern "win64" fn increment_igt(registers: *mut Registers, _:usize)
 // A second patch, "FPS history" below, is required in addition to this one to ensure accuracy.
 unsafe extern "win64" fn fps(registers: *mut Registers, _:usize)
 {
-    if ER_FPS_PATCH_ENABLED
+    unsafe
     {
-        let ptr_flipper = (*registers).rbx as *const u8; // Flipper struct - Contains all the stuff we need
+        if ER_FPS_PATCH_ENABLED
+        {
+            let ptr_flipper = (*registers).rbx as *const u8; // Flipper struct - Contains all the stuff we need
 
-        let ptr_target_frame_delta = ptr_flipper.offset(FPS_OFFSETS.target_frame_delta) as *mut f32; // Target frame delta - Set in a switch/case at the start
-        let ptr_timestamp_previous = ptr_flipper.offset(FPS_OFFSETS.timestamp_previous) as *mut u64; // Previous frames timestamp
-        let ptr_timestamp_current = ptr_flipper.offset(FPS_OFFSETS.timestamp_current) as *mut u64; // Current frames timestamp
-        let ptr_frame_delta = ptr_flipper.offset(FPS_OFFSETS.frame_delta) as *mut f32; // Current frames frame delta
+            let ptr_target_frame_delta = ptr_flipper.offset(FPS_OFFSETS.target_frame_delta) as *mut f32; // Target frame delta - Set in a switch/case at the start
+            let ptr_timestamp_previous = ptr_flipper.offset(FPS_OFFSETS.timestamp_previous) as *mut u64; // Previous frames timestamp
+            let ptr_timestamp_current = ptr_flipper.offset(FPS_OFFSETS.timestamp_current) as *mut u64; // Current frames timestamp
+            let ptr_frame_delta = ptr_flipper.offset(FPS_OFFSETS.frame_delta) as *mut f32; // Current frames frame delta
 
-        // Read target frame data, the current timestamp and then calculate the timestamp diff at stable FPS
-        let target_frame_delta = std::ptr::read_volatile(ptr_target_frame_delta);
-        let timestamp_current = std::ptr::read_volatile(ptr_timestamp_current);
-        let timestamp_diff = (target_frame_delta * 10000000.0) as i32;
+            // Read target frame data, the current timestamp and then calculate the timestamp diff at stable FPS
+            let target_frame_delta = std::ptr::read_volatile(ptr_target_frame_delta);
+            let timestamp_current = std::ptr::read_volatile(ptr_timestamp_current);
+            let timestamp_diff = (target_frame_delta * 10000000.0) as i32;
 
-        // Calculate the previous timestamp, as well as the frame delta
-        let timestamp_previous = timestamp_current - (timestamp_diff as u64);
-        let frame_delta = (timestamp_diff as f32) / 10000000.0;
+            // Calculate the previous timestamp, as well as the frame delta
+            let timestamp_previous = timestamp_current - (timestamp_diff as u64);
+            let frame_delta = (timestamp_diff as f32) / 10000000.0;
 
-        // Write values back
-        std::ptr::write_volatile(ptr_timestamp_previous, timestamp_previous);
-        std::ptr::write_volatile(ptr_frame_delta, frame_delta);
+            // Write values back
+            std::ptr::write_volatile(ptr_timestamp_previous, timestamp_previous);
+            std::ptr::write_volatile(ptr_frame_delta, frame_delta);
+        }
     }
 }
 
@@ -208,15 +248,18 @@ unsafe extern "win64" fn fps(registers: *mut Registers, _:usize)
 // This gets stored in an array with 32 elements, possibly for calculating FPS averages.
 unsafe extern "win64" fn fps_history(registers: *mut Registers, _:usize)
 {
-    if ER_FPS_PATCH_ENABLED
+    unsafe
     {
-        let ptr_flipper = (*registers).rbx as *const u8; // Flipper struct - Contains all the stuff we need
+        if ER_FPS_PATCH_ENABLED
+        {
+            let ptr_flipper = (*registers).rbx as *const u8; // Flipper struct - Contains all the stuff we need
 
-        let ptr_target_frame_delta = ptr_flipper.offset(FPS_OFFSETS.target_frame_delta) as *mut f32; // Target frame delta - Set in a switch/case at the start
+            let ptr_target_frame_delta = ptr_flipper.offset(FPS_OFFSETS.target_frame_delta) as *mut f32; // Target frame delta - Set in a switch/case at the start
 
-        // Read the target frame delta and write back the calculated frame delta timestamp
-        let target_frame_delta = std::ptr::read_volatile(ptr_target_frame_delta);
-        (*registers).rax = (target_frame_delta * 10000000.0) as u64;
+            // Read the target frame delta and write back the calculated frame delta timestamp
+            let target_frame_delta = std::ptr::read_volatile(ptr_target_frame_delta);
+            (*registers).rax = (target_frame_delta * 10000000.0) as u64;
+        }
     }
 }
 
@@ -226,21 +269,24 @@ unsafe extern "win64" fn fps_history(registers: *mut Registers, _:usize)
 // This does not allow you to go above the stock FPS limit. It is purely a QoL patch to improve glitch consistency, not an FPS unlocker.
 unsafe extern "win64" fn fps_custom_limit(registers: *mut Registers, _:usize)
 {
-    if ER_FPS_PATCH_ENABLED && ER_FPS_CUSTOM_LIMIT > 0.0f32
+    unsafe
     {
-        let ptr_flipper = (*registers).rbx as *const u8; // Flipper struct - Contains all the stuff we need
-
-        let ptr_target_frame_delta = ptr_flipper.offset(FPS_OFFSETS.target_frame_delta) as *mut f32; // Target frame delta - Set in a switch/case at the start
-
-        // Read the stock target frame delta and calculate the custom target frame delta
-        let target_frame_delta = std::ptr::read_volatile(ptr_target_frame_delta);
-        let custom_target_frame_delta = 1.0f32 / ER_FPS_CUSTOM_LIMIT;
-
-        // Make sure the custom target frame delta is higher than the stock one, in order to avoid going above the stock FPS limit
-        if custom_target_frame_delta > target_frame_delta
+        if ER_FPS_PATCH_ENABLED && ER_FPS_CUSTOM_LIMIT > 0.0f32
         {
-            // Write values back
-            std::ptr::write_volatile(ptr_target_frame_delta, custom_target_frame_delta);
+            let ptr_flipper = (*registers).rbx as *const u8; // Flipper struct - Contains all the stuff we need
+
+            let ptr_target_frame_delta = ptr_flipper.offset(FPS_OFFSETS.target_frame_delta) as *mut f32; // Target frame delta - Set in a switch/case at the start
+
+            // Read the stock target frame delta and calculate the custom target frame delta
+            let target_frame_delta = std::ptr::read_volatile(ptr_target_frame_delta);
+            let custom_target_frame_delta = 1.0f32 / ER_FPS_CUSTOM_LIMIT;
+
+            // Make sure the custom target frame delta is higher than the stock one, in order to avoid going above the stock FPS limit
+            if custom_target_frame_delta > target_frame_delta
+            {
+                // Write values back
+                std::ptr::write_volatile(ptr_target_frame_delta, custom_target_frame_delta);
+            }
         }
     }
 }
@@ -248,12 +294,33 @@ unsafe extern "win64" fn fps_custom_limit(registers: *mut Registers, _:usize)
 // Frame advance patch
 unsafe extern "win64" fn frame_advance(_registers: *mut Registers, _:usize)
 {
-    if ER_FRAME_ADVANCE_ENABLED
+    unsafe
     {
-        ER_FRAME_RUNNING = false;
+        if ER_FRAME_ADVANCE_ENABLED
+        {
+            ER_FRAME_RUNNING = false;
 
-        while !ER_FRAME_RUNNING && ER_FRAME_ADVANCE_ENABLED {
-            thread::sleep(Duration::from_micros(10));
+            while !ER_FRAME_RUNNING && ER_FRAME_ADVANCE_ENABLED {
+                thread::sleep(Duration::from_micros(10));
+            }
         }
     }
 }
+
+
+pub unsafe extern "win64" fn xinput_fn(registers: *mut Registers, orig_func_ptr: usize, _: usize) -> usize { unsafe {
+    
+    let dw_user_index = (*registers).rcx as u32;
+    let p_state = (*registers).rdx as *mut XINPUT_STATE;
+
+    if !ER_XINPUT_PATCH_ENABLED {
+        let orig_func: XInputGetState = mem::transmute(orig_func_ptr);
+        return orig_func(dw_user_index, p_state) as usize;
+    }
+
+    (*p_state) = ER_XINPUT_STATE;
+
+    // ERROR_SUCCESS = 0x0
+    // ERROR_DEVICE_NOT_CONNECTED = 0x48F
+    return 0x0;
+}}
